Implement `Backend.get_session` to retrieve scoped session
The scoped session is an instance of SqlAlchemy's `Session` class that
is used by the query builder to connect to the database, for both the
SqlAlchemy database backend as well as for Django. Both database
backends need to maintain their own scoped session factory which can be
called to get a session instance.

Certain applications need access to the session. For example,
applications that run AiiDA in a threaded way, such as a REST API server
need to manually close the session after the query has finished because
this is not done automatically when the thread ends. The associated
database connection remains open causing an eventual timeout when a new
request comes in. The method `Backend.get_session` provides an official
API to access the global scoped session instance which can then be
closed.

Additionally, a lot of code that was duplicated across the two
implementations of the `QueryBuilder` for the two database backends has
been moved to the abstract `BackendQueryBuilder`. Normally this code
does indeed belong in the implementations but since the current
implementation for both backends is based on SqlAlchemy they are both
nearly identical. When in the future a new backend is implemented that
does not use SqlAlchemy the current code can be factored out to a
specific `SqlAlchemyQueryBuilder` that can be used for both database
backends.
