+++ aiida/orm/implementation/querybuilder.py
@@ -7,8 +7,21 @@
 # For further information on the license, see the LICENSE.txt file        #
 # For further information please visit http://www.aiida.net               #
 ###########################################################################
-"""Backend query implementation classes"""
+"""Abstract `QueryBuilder` definition.
+
+Note that this abstract class actually contains parts of the implementation, which are tightly coupled to SqlAlchemy.
+This is done because currently, both database backend implementations, both Django and SqlAlchemy, directly use the
+SqlAlchemy library to implement the query builder. If there ever is another database backend to be implemented that does
+not go through SqlAlchemy, this class will have to be refactored. The SqlAlchemy specific implementations should most
+likely be moved to a `SqlAlchemyBasedQueryBuilder` class and restore this abstract class to being a pure agnostic one.
+"""
 import abc
+import uuid
+
+# pylint: disable=no-name-in-module, import-error
+from sqlalchemy_utils.types.choice import Choice
+from sqlalchemy.types import Integer, Float, Boolean, DateTime
+from sqlalchemy.dialects.postgresql import JSONB
 
 from aiida.common import exceptions
 from aiida.common.lang import abstractclassmethod, type_check
@@ -100,11 +113,11 @@ class BackendQueryBuilder:
         from aiida.orm import Node
         return Node
 
-    @abc.abstractmethod
     def get_session(self):
         """
-        :returns: a valid session, an instance of sqlalchemy.orm.session.Session
+        :returns: a valid session, an instance of :class:`sqlalchemy.orm.session.Session`
         """
+        return self._backend.get_session()
 
     @abc.abstractmethod
     def modify_expansions(self, alias, expansions):
@@ -210,23 +223,51 @@ class BackendQueryBuilder:
             raise InputValidationError('Unknown operator {} for filters on columns'.format(operator))
         return expr
 
-    @abc.abstractmethod
     def get_projectable_attribute(self, alias, column_name, attrpath, cast=None, **kwargs):
-        pass
+        """
+        :returns: An attribute store in a JSON field of the give column
+        """
+        # pylint: disable=unused-argument
+        entity = self.get_column(column_name, alias)[attrpath]
+        if cast is None:
+            pass
+        elif cast == 'f':
+            entity = entity.astext.cast(Float)
+        elif cast == 'i':
+            entity = entity.astext.cast(Integer)
+        elif cast == 'b':
+            entity = entity.astext.cast(Boolean)
+        elif cast == 't':
+            entity = entity.astext
+        elif cast == 'j':
+            entity = entity.astext.cast(JSONB)
+        elif cast == 'd':
+            entity = entity.astext.cast(DateTime)
+        else:
+            raise InputValidationError('Unkown casting key {}'.format(cast))
+        return entity
 
-    @abc.abstractmethod
-    def get_aiida_res(self, key, res):
+    def get_aiida_res(self, res):
         """
         Some instance returned by ORM (django or SA) need to be converted
-        to Aiida instances (eg nodes)
+        to AiiDA instances (eg nodes). Choice (sqlalchemy_utils)
+        will return their value
 
-        :param key: the key that this entry would be returned with
         :param res: the result returned by the query
 
         :returns: an aiida-compatible instance
         """
+        if isinstance(res, Choice):
+            return res.value
 
-    @abc.abstractmethod
+        if isinstance(res, uuid.UUID):
+            return str(res)
+
+        try:
+            return self._backend.get_backend_entity(res)
+        except TypeError:
+            return res
+
     def yield_per(self, query, batch_size):
         """
         :param int batch_size: Number of rows to yield per step
@@ -235,32 +276,117 @@ class BackendQueryBuilder:
 
         :returns: a generator
         """
+        try:
+            return query.yield_per(batch_size)
+        except Exception:
+            self.get_session().close()
+            raise
 
-    @abc.abstractmethod
     def count(self, query):
         """
         :returns: the number of results
         """
+        try:
+            return query.count()
+        except Exception:
+            self.get_session().close()
+            raise
 
-    @abc.abstractmethod
     def first(self, query):
         """
         Executes query in the backend asking for one instance.
 
         :returns: One row of aiida results
         """
+        try:
+            return query.first()
+        except Exception:
+            self.get_session().close()
+            raise
 
-    @abc.abstractmethod
     def iterall(self, query, batch_size, tag_to_index_dict):
         """
         :return: An iterator over all the results of a list of lists.
         """
+        try:
+            if not tag_to_index_dict:
+                raise Exception('Got an empty dictionary: {}'.format(tag_to_index_dict))
 
-    @abc.abstractmethod
+            results = query.yield_per(batch_size)
+
+            if len(tag_to_index_dict) == 1:
+                # Sqlalchemy, for some strange reason, does not return a list of lsits
+                # if you have provided an ormclass
+
+                if list(tag_to_index_dict.values()) == ['*']:
+                    for rowitem in results:
+                        yield [self.get_aiida_res(rowitem)]
+                else:
+                    for rowitem, in results:
+                        yield [self.get_aiida_res(rowitem)]
+            elif len(tag_to_index_dict) > 1:
+                for resultrow in results:
+                    yield [self.get_aiida_res(rowitem) for colindex, rowitem in enumerate(resultrow)]
+            else:
+                raise ValueError('Got an empty dictionary')
+        except Exception:
+            self.get_session().close()
+            raise
+
     def iterdict(self, query, batch_size, tag_to_projected_properties_dict, tag_to_alias_map):
         """
         :returns: An iterator over all the results of a list of dictionaries.
         """
+        try:
+            nr_items = sum(len(v) for v in tag_to_projected_properties_dict.values())
+
+            if not nr_items:
+                raise ValueError('Got an empty dictionary')
+
+            results = query.yield_per(batch_size)
+            if nr_items > 1:
+                for this_result in results:
+                    yield {
+                        tag: {
+                            self.get_corresponding_property(
+                                self.get_table_name(tag_to_alias_map[tag]), attrkey, self.inner_to_outer_schema
+                            ): self.get_aiida_res(this_result[index_in_sql_result])
+                            for attrkey, index_in_sql_result in projected_entities_dict.items()
+                        } for tag, projected_entities_dict in tag_to_projected_properties_dict.items()
+                    }
+            elif nr_items == 1:
+                # I this case, sql returns a  list, where each listitem is the result
+                # for one row. Here I am converting it to a list of lists (of length 1)
+                if [v for entityd in tag_to_projected_properties_dict.values() for v in entityd.keys()] == ['*']:
+                    for this_result in results:
+                        yield {
+                            tag: {
+                                self.get_corresponding_property(
+                                    self.get_table_name(tag_to_alias_map[tag]), attrkey, self.inner_to_outer_schema
+                                ): self.get_aiida_res(this_result)
+                                for attrkey, position in projected_entities_dict.items()
+                            } for tag, projected_entities_dict in tag_to_projected_properties_dict.items()
+                        }
+                else:
+                    for this_result, in results:
+                        yield {
+                            tag: {
+                                self.get_corresponding_property(
+                                    self.get_table_name(tag_to_alias_map[tag]), attrkey, self.inner_to_outer_schema
+                                ): self.get_aiida_res(this_result)
+                                for attrkey, position in projected_entities_dict.items()
+                            } for tag, projected_entities_dict in tag_to_projected_properties_dict.items()
+                        }
+            else:
+                raise ValueError('Got an empty dictionary')
+        except Exception:
+            self.get_session().close()
+            raise
+
+    @staticmethod
+    @abstractclassmethod
+    def get_table_name(aliased_class):
+        """Returns the table name given an Aliased class."""
 
     @abc.abstractmethod
     def get_column_names(self, alias):
